requirement.txt
fastapi
uvicorn[standard]
python-socketio[asyncio_server]
keras
tensorflow
numpy
pillow
opencv-python
python-multipart
aiofiles
